CIF-2827: Refactoring
