+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/experiencefragment/CommerceExperienceFragmentContainer.java
@@ -13,30 +13,24 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-package com.adobe.cq.commerce.core.components.internal.models.v1.common;
+package com.adobe.cq.commerce.core.components.models.experiencefragment;
 
 import org.apache.sling.api.resource.Resource;
 
-import com.adobe.cq.commerce.core.components.models.common.CommerceExperienceFragmentContainer;
+import com.adobe.cq.wcm.core.components.models.Component;
 
-public class CommerceExperienceFragmentContainerImpl implements CommerceExperienceFragmentContainer {
+/**
+ * Interface for an Experience Fragment container to be used in a
+ * {@code ProductList}
+ */
+public interface CommerceExperienceFragmentContainer extends Component {
 
-    private Resource renderResource;
-    private String cssClassName;
+    String getCssClassName();
 
-    public CommerceExperienceFragmentContainerImpl(Resource renderResource, String cssClassName) {
-        this.renderResource = renderResource;
-        this.cssClassName = cssClassName;
-    }
-
-    @Override
-    public Resource getRenderResource() {
-        return renderResource;
-    }
-
-    @Override
-    public String getCssClassName() {
-        return cssClassName;
-    }
-
+    /**
+     * Returns the ExperienceFragment resource to be rendered.
+     *
+     * @return the ExperienceFragment resource to be rendered
+     */
+    Resource getRenderResource();
 }
