+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/experiencefragment/package-info.java
@@ -13,7 +13,7 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-@Version("1.12.0")
+@Version("1.13.0")
 package com.adobe.cq.commerce.core.components.models.experiencefragment;
 
 import org.osgi.annotation.versioning.Version;
