+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/productlist/ProductList.java
@@ -22,7 +22,7 @@ import javax.annotation.Nullable;
 
 import org.osgi.annotation.versioning.ConsumerType;
 
-import com.adobe.cq.commerce.core.components.models.common.CommerceExperienceFragmentContainer;
+import com.adobe.cq.commerce.core.components.models.experiencefragment.CommerceExperienceFragmentContainer;
 import com.adobe.cq.commerce.core.components.models.page.PageMetadata;
 import com.adobe.cq.commerce.core.components.models.productcollection.ProductCollection;
 import com.adobe.cq.commerce.core.components.models.retriever.AbstractCategoryRetriever;
